+++ R/namespace.R
@@ -1,2 +1,3 @@
-#' @import err checkr mcmcr
+#' @import chk universals extras
+#' @importFrom mcmcr as.mcmcarray as.mcmcr as.mcmcrs bind_parameters
 NULL