+++ NEWS.md
@@ -1,3 +1,11 @@
+# mcmcderive 0.1.0
+
+- Added `mcmc_derive()` for nlist, nlists and mcmc objects
+- Added `primary = FALSE` argument to `mcmc_derive()` to include original parameters as is.
+- Adopted Tidyverse style for error messages.
+- Moved extra functions to `extras` package.
+- Replaced `checkr` and `err` dependencies for `chk`.
+
 # mcmcderive 0.0.1
 
 * Optimized `mcmc_derive()`.